requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
bcrypt==4.1.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
email-validator==2.1.0
Pillow==10.1.0
google-auth==2.27.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
